realizando o exercicio04-Rio

+++ 4-semestre/programacao-orientada-a-objetos/listas/lista01/src/main/java/br/edu/fatecfranca/exercicio04/Rio.java
@@ -2,7 +2,7 @@ package br.edu.fatecfranca.exercicio04;
 
 public class Rio {
     public String nome;
-    public float nivel;
+    public double nivel;
     public boolean poluido;
 
     public Rio(String nome, float nivel, boolean poluido) {
